[Backport try #2] Add Xilinx BSCANE2 tap
Alternative tap implementation for Xilinx devices which are not easily
accessible by means of a dedicated JTAG header. This implementation
should also be portable amongst 7-series and Ultrascale+ devices.

The commit also contains a testbench for the regular JTAG TAP and the
BSCANE2 TAP (same testbnech really). Both testbenches pass and hopefully
give confidence that I didn't introduce a regression.